[refactor] add an execute file in native code
this commit add an execute.cpp file there execute another process in a new session
it also contain makefile used to compile the file.
In order to simplify the workflow to users the src/out directory will contain binary
executables.
When installing this module 'node src/setup.js' will be executed and it will symlink
the correct file depending on the OS to lib/executables